requirements.txt
pandas
tqdm
regex 
transformers 
datasets 
sentencepiece
torch
sacremoses
nltk
sentence_transformers
faiss-cpu